% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lalaland_palettes.R
\docType{data}
\name{lalaland_palettes}
\alias{lalaland_palettes}
\title{List of "La La Land" palettes}
\format{A list with 10 named elements.}
\usage{
lalaland_palettes
}
\description{
Hexadecimal codes for all 10 palettes.
}
\details{
Created using stills from "La La Land", image color pickers, and \url{https://projects.susielu.com/viz-palette}.
}
\keyword{datasets}
