% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lalaland_palettes.R
\name{lll_palette}
\alias{lll_palette}
\title{Generates color palettes from "La La Land"}
\usage{
lll_palette(name, n)
}
\arguments{
\item{name}{Name of the desired color palette. Choices:
\code{CityOfStars1}, \code{CityOfStars2}, \code{LightHouseCafe},
\code{Sebs}, \code{Summer}, \code{Epilogue1}, \code{California},
\code{LovelyNight}, \code{Epilogue2}, \code{Traffic}.}

\item{n}{Vector of colors desired. Colors can be selected, and
combinations viewed, using \code{show_lll_palette}.}
}
\value{
A vector of colors, in the form of hexadecimal codes.
}
\description{
Hexadecimal codes for a selection of color palettes inspired by the 2016 film "La La Land".
}
\examples{
lll_palette("CityOfStars1")
lll_palette("Summer", c(1,3,5))

## In a ggplot2 context:
library(ggplot2)
ggplot(mtcars, aes(x = disp, y = qsec, color = as.factor(carb))) +
geom_point() +
scale_color_manual(values = lll_palette("Epilogue1")) +
theme_minimal()

}
