% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lalaland_palettes.R
\name{show_lll_palette}
\alias{show_lll_palette}
\title{Displays color palettes from "La La Land"}
\usage{
show_lll_palette(name, n)
}
\arguments{
\item{name}{Name of the desired color palette. Choices:
\code{CityOfStars1}, \code{CityOfStars2}, \code{LightHouseCafe},
\code{Sebs}, \code{Summer}, \code{Epilogue1}, \code{California},
\code{LovelyNight}, \code{Epilogue2}, \code{Traffic}.}

\item{n}{Vector of colors to be viewed in palette form.}
}
\description{
Prints images that depict color palettes inspired by the 2016 film "La La Land".
}
